dashboard/requirements.txt
# Kafka
kafka-python==2.0.2
websocket-client==1.6.1

# Database
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9

# API requests & environment variables
requests==2.32.3
python-dotenv==1.0.1

# Data manipulation
pandas==2.2.2
numpy==1.26.4

# Visualization & dashboard
streamlit==1.38.0
plotly==5.24.0
altair==5.3.0

# Machine learning
scikit-learn==1.5.1

# LLM integration
google-genai

seaborn==0.12.2
matplotlib==3.8.1
